refactor(font/glyph)!: reorganize export paths
