+++ crates/monoxide-font/src/font.rs
@@ -39,7 +39,7 @@ pub fn make_font() -> Result<FontContext, ()> {
 
     let glyphs = [
         (' ', glyph::space(&cx)),
-        ('A', glyph::a::a_cap(&cx)),
+        ('A', glyph::a_cap(&cx)),
         ('a', glyph::a(&cx)),
         ('b', glyph::b(&cx)),
         ('c', glyph::c(&cx)),
@@ -49,13 +49,13 @@ pub fn make_font() -> Result<FontContext, ()> {
         ('g', glyph::g(&cx)),
         ('h', glyph::h(&cx)),
         ('i', glyph::i(&cx)),
-        ('J', glyph::j::j_cap(&cx)),
+        ('J', glyph::j_cap(&cx)),
         ('j', glyph::j(&cx)),
         ('k', glyph::k(&cx)),
         ('l', glyph::l(&cx)),
         ('m', glyph::m(&cx)),
         ('n', glyph::n(&cx)),
-        ('O', glyph::o::o_cap(&cx)),
+        ('O', glyph::o_cap(&cx)),
         ('o', glyph::o(&cx)),
         ('p', glyph::p(&cx)),
         ('q', glyph::q(&cx)),
