+++ crates/monoxide-font/src/font/glyph.rs
@@ -1,4 +1,4 @@
-pub mod a;
+mod a;
 mod b;
 mod c;
 mod d;
@@ -7,19 +7,18 @@ mod f;
 mod g;
 mod h;
 mod i;
-pub mod j;
+mod j;
 mod k;
 mod l;
 mod m;
 mod n;
-pub mod o;
+mod o;
 mod p;
 mod q;
 mod r;
 mod s;
-mod space;
+mod sym;
 mod t;
-mod tofu;
 mod u;
 mod v;
 mod w;
@@ -28,8 +27,32 @@ mod y;
 mod z;
 
 pub use self::{
-    a::a, b::b, c::c, d::d, e::e, f::f, g::g, h::h, i::i, j::j, k::k, l::l, m::m, n::n, o::o, p::p,
-    q::q, r::r, s::s, space::space, t::t, tofu::tofu, u::u, v::v, w::w, x::x, y::y, z::z,
+    a::{a, a_cap},
+    b::b,
+    c::c,
+    d::d,
+    e::e,
+    f::f,
+    g::g,
+    h::h,
+    i::i,
+    j::{j, j_cap},
+    k::k,
+    l::l,
+    m::m,
+    n::n,
+    o::{o, o_cap},
+    p::p,
+    q::q,
+    r::r,
+    s::s,
+    sym::{space, tofu},
+    t::t,
+    u::u,
+    v::v,
+    w::w,
+    x::x,
+    y::y,
+    z::z,
 };
-
-use super::InputContext;
+use crate::InputContext;
