+++ crates/monoxide-font/src/font/glyph/d.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use monoxide_script::prelude::*;
 
-use super::InputContext;
+use crate::InputContext;
 use crate::font::{dir::Alignment, glyph::o::OShape, settings::FontParamSettings, shape::Rect};
 
 pub fn d(cx: &InputContext) -> Glyph {
