--- crates/monoxide-font/src/font/glyph/space.rs
@@ -1,7 +1,0 @@
-use monoxide_script::ast::Glyph;
-
-use super::InputContext;
-
-pub fn space(_cx: &InputContext) -> Glyph {
-    Glyph::builder().build()
-}
