+++ crates/monoxide-font/src/font/glyph/v.rs
@@ -1,6 +1,6 @@
 use monoxide_script::prelude::*;
 
-use super::InputContext;
+use crate::InputContext;
 use crate::font::dir::{Alignment, Dir};
 
 pub fn v(cx: &InputContext) -> Glyph {
