+++ crates/monoxide-font/src/font/glyph/w.rs
@@ -2,7 +2,7 @@ use std::{ops::Range, sync::Arc};
 
 use monoxide_script::prelude::*;
 
-use super::InputContext;
+use crate::InputContext;
 use crate::font::{
     dir::{Alignment, Dir},
     math::mix,
